+++ cider-repl.el
@@ -92,15 +92,6 @@ change the setting's value."
   :type 'boolean
   :group 'cider-repl)
 
-(defcustom cider-repl-print-length nil
-  "Non-nil means limit the number of objects printed in REPL to this value.
-This is implemented by setting the value of the Clojure var
-*print-length*.  The `cider-repl-toggle-print-length-limiting'
-command can be used to interactively change whether this setting
-is enforced or not."
-  :type 'integer
-  :group 'cider-repl)
-
 (defcustom cider-repl-use-clojure-font-lock nil
   "Non-nil means to use Clojure mode font-locking for input and result.
 Nil means that `cider-repl-input-face' and `cider-repl-result-face'
@@ -212,8 +203,6 @@ Insert a banner, unless NOPROMPT is non-nil."
   (with-current-buffer buffer
     (unless (eq major-mode 'cider-repl-mode)
       (cider-repl-mode))
-    (when cider-repl-print-length
-      (cider-repl-set-print-length cider-repl-print-length))
     (cider-repl-reset-markers)
     ;; honor :init-ns from lein's :repl-options on startup
     (setq nrepl-buffer-ns (cider-eval-and-get-value "(str *ns*)"))
@@ -633,23 +622,6 @@ text property `cider-old-input'."
   (setq cider-repl-use-pretty-printing (not cider-repl-use-pretty-printing))
   (message "Pretty printing in nREPL %s."
            (if cider-repl-use-pretty-printing "enabled" "disabled")))
-
-(defun cider-repl-set-print-length (length)
-  "Set the clojure var *print-length* to LENGTH."
-  (let* ((form (format "(set! *print-length* %d)"
-                       length)))
-    (cider-eval-sync form)))
-
-(defun cider-repl-toggle-print-length-limiting ()
-  "Toggle the enforcement of `cider-repl-print-length'."
-  (interactive)
-  (when (integerp cider-repl-print-length)
-    (let* ((form (format "(set! *print-length* (if *print-length* nil %d))"
-                         cider-repl-print-length))
-           (print-length (cider-eval-and-get-value form)))
-      (if print-length
-          (message "*print-length* limited to %d" print-length)
-        (message "*print-length* unlimited")))))
 
 (defvar cider-repl-clear-buffer-hook)